requirements.txt
opencv-python
numpy
bluepy
deepface
python-miio>=0.5.12
python-dotenv
